+++ src/Oliverde8/Component/PhpEtl/ChainOperation/ChainSplitOperation.php
@@ -35,7 +35,7 @@ class ChainSplitOperation extends AbstractChainOperation implements DataChainOpe
     public function processData(DataItemInterface $item, ExecutionContext $context): ItemInterface
     {
         foreach ($this->chainProcessors as $chainProcessor) {
-            $chainProcessor->processItem($item, 0,  $context);
+            $chainProcessor->processItemWithChain($item, 0,  $context);
         }
 
         // Nothing to process.
@@ -45,7 +45,7 @@ class ChainSplitOperation extends AbstractChainOperation implements DataChainOpe
     public function processStop(StopItem $item, ExecutionContext $context): ItemInterface
     {
         foreach ($this->chainProcessors as $chainProcessor) {
-            $result = $chainProcessor->processItem($item, 0,  $context);
+            $result = $chainProcessor->processItemWithChain($item, 0,  $context);
 
             if ($result !== $item) {
                 // Return a new stop item in order to continue flushing out data with stop items.