+++ src/Oliverde8/Component/PhpEtl/Tests/ChainProcessorTest.php
@@ -8,6 +8,7 @@
 
 namespace Oliverde8\Component\PhpEtl\Tests;
 
+use Oliverde8\Component\PhpEtl\Tests\Item\TestAsyncItem;
 use Oliverde8\Component\PhpEtl\ChainOperation\Grouping\SimpleGroupingOperation;
 use Oliverde8\Component\PhpEtl\ChainOperation\Transformer\CallbackTransformerOperation;
 use Oliverde8\Component\PhpEtl\ChainProcessor;
@@ -178,6 +179,90 @@ class ChainProcessorTest extends TestCase
 
         $this->assertEquals([1, 2], $values1);
         $this->assertEquals([1, 3, 4, 5, 2, 3, 4, 5], $values2);
+    }
+
+    public function testAsyncItems()
+    {
+        $results = [];
+        $callNum = 0;
+
+        $mock = new CallbackTransformerOperation(function (ItemInterface $item) use (&$callNum) {
+            $callNum++;
+            if ($callNum == 1) {
+                return new TestAsyncItem(new DataItem('I am slow'), 2);
+            } else {
+                return new TestAsyncItem(new DataItem('I am fast'), 1);
+            }
+        });
+        $mockEnd = new CallbackTransformerOperation(function (ItemInterface $item) use (&$results) {
+            $results[] = $item->getData();
+            return $item;
+        });
+
+        $chainProcessor = new ChainProcessor(
+            ['mock' => $mock, 'mocked_end' => $mockEnd],
+            new ExecutionContextFactory()
+        );
+        $chainProcessor->process(new \ArrayIterator([1, 2]), ['toto']);
+
+        // Order of items changes
+        $this->assertEquals(["I am fast", "I am slow"], $results);
+    }
+
+    public function testMaxAsyncItems()
+    {
+        $results = [];
+        $callNum = 0;
+
+        $mock = new CallbackTransformerOperation(function (ItemInterface $item) use (&$callNum) {
+            $callNum++;
+            if ($callNum == 1) {
+                return new TestAsyncItem(new DataItem('I am slow1'), 2);
+            } elseif ($callNum == 2) {
+                return new TestAsyncItem(new DataItem('I am slow2'), 3);
+            } else {
+                return new TestAsyncItem(new DataItem('I am speed'), 1);
+            }
+        });
+        $mockEnd = new CallbackTransformerOperation(function (ItemInterface $item) use (&$results) {
+            $results[] = $item->getData();
+            return $item;
+        });
+
+        $chainProcessor = new ChainProcessor(
+            ['mock' => $mock, 'mocked_end' => $mockEnd],
+            new ExecutionContextFactory(),
+            2
+        );
+        $chainProcessor->process(new \ArrayIterator([1, 2, 3]), ['toto']);
+
+        // Speed can't catch up with slow1, because the queue is full.
+        $this->assertEquals(["I am slow1", "I am speed", "I am slow2"], $results);
+    }
+
+    public function testAsyncItemsWithMixItem()
+    {
+        $results = [];
+
+        $mock = new CallbackTransformerOperation(function (ItemInterface $item) use (&$callNum) {
+            return new MixItem([
+                new TestAsyncItem(new DataItem('I am slow'), 2),
+                new TestAsyncItem(new DataItem('I am fast'), 1),
+            ]);
+        });
+        $mockEnd = new CallbackTransformerOperation(function (ItemInterface $item) use (&$results) {
+            $results[] = $item->getData();
+            return $item;
+        });
+
+        $chainProcessor = new ChainProcessor(
+            ['mock' => $mock, 'mocked_end' => $mockEnd],
+            new ExecutionContextFactory()
+        );
+        $chainProcessor->process(new \ArrayIterator([1]), ['toto']);
+
+        // Order of items changes
+        $this->assertEquals(["I am fast", "I am slow"], $results);
     }
 
     public function testException()